refactor: update education section format and use full month names in dates

+++ generate-resume.js
@@ -16,7 +16,7 @@ Handlebars.registerHelper('formatDate', function(dateString) {
   
   try {
     const date = new Date(dateString);
-    return date.toLocaleDateString('en-US', { month: 'short', year: 'numeric' });
+    return date.toLocaleDateString('en-US', { month: 'long', year: 'numeric' });
   } catch (e) {
     return dateString; // Return the original string if parsing fails
   }
